Fix checking compile output

+++ test.sh
@@ -12,18 +12,25 @@ passed=0
 failed=0
 
 for xib in $xibs_to_test ; do
-	python3 -m ibtool --compile "$test_out" "$xib"
+	compile_output=$(python3 -m ibtool --compile "$test_out" "$xib" 2>&1)
 
-	output=$(python3 -m ibtool --compare "${xib/xib/nib}" "$test_out" 2>&1)
-	if [[ $? == 0 ]] ; then
-		passed=$((passed + 1))
+	if [ -r "$test_out" ] ; then
+		output=$(python3 -m ibtool --compare "${xib/xib/nib}" "$test_out" 2>&1)
+		if [[ $? == 0 ]] ; then
+			passed=$((passed + 1))
+		else
+			failed=$((failed + 1))
+			echo "FAIL: $xib"
+			echo "$output"
+			echo
+		fi
+		rm -f "$test_out"
 	else
 		failed=$((failed + 1))
 		echo "FAIL: $xib"
-		echo "$output"
+		echo "$compile_output"
 		echo
 	fi
-	rm -f "$test_out"
 done
 
 echo "Passed: $passed  Failed: $failed"
